+++ Dozer/StatusIconClasses/NormalStatusIcon.swift
@@ -28,7 +28,7 @@ class NormalStatusIcon: HelperstatusIcon {
         case .leftMouseDown:
             DozerIcons.shared.toggle()
         case .rightMouseDown:
-            appDelegate.preferencesWindowController.show(preferencePane: .general)
+            AppDelegate.shared.preferencesWindowController.show(preferencePane: .general)
         default:
             break
         }